refactor:uploader common logic consolidated into BaseUploader
